requirements.txt
boto3==1.34.146
cohere==5.6.2
pinecone==4.0.0
pyairtable==2.3.3
sumy==0.11.0
telethon==1.36.0
tenacity==8.5.0
tokenizers==0.15.2 ; platform_machine == 'x86_64'
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
google-generativeai==0.3.2